+++ src/data/components/table/DataCells.tsx
@@ -6,7 +6,6 @@ import CvCell from './cells/CvCell'
 import ProjectStatusCell from './cells/ProjectStatusCell'
 import CustomerStatusCell from './cells/CustomerStatusCell'
 import CheckBoxHeaderCell from './cells/CheckBoxHeaderCell'
-import CenteredHeaderCell from './cells/CenteredHeaderCell'
 
 const EducationCell = ({ data }: { data: string | null }) => (
   <CharacterLimitBox text={data || '-'} />
@@ -19,5 +18,4 @@ export {
   ProjectStatusCell,
   CustomerStatusCell,
   CheckBoxHeaderCell,
-  CenteredHeaderCell,
 }
